fix(load_data): move to separate file to avoid circular dependency
